Feature/binding gpio without library (#43)
Implementation of the bindings for the Raspberry Pi 4b without the rppal library in a user space different from the physical address space, so that we do not put it in our trusted computing base. Moreover we implemented it such that is uses the minimum amount of abstractions so that it reduces complexity and the amount of functions/operations/codes putted in our trusted computing base.